+++ app/layout.tsx
@@ -13,8 +13,8 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Job Application Traccker",
+  description: "Developed by Reza Brilliansah on Next.js",
 };
 
 export default function RootLayout({
